+++ mjolnir/input/read_external_interaction.hpp
@@ -2,8 +2,8 @@
 #define MJOLNIR_INPUT_READ_EXTERNAL_INTERACTION_HPP
 #include <extlib/toml/toml.hpp>
 #include <mjolnir/forcefield/AFMFit/AFMFitInteraction.hpp>
-#include <mjolnir/interaction/external/ExternalDistanceInteraction.hpp>
-#include <mjolnir/interaction/external/PositionRestraintInteraction.hpp>
+#include <mjolnir/forcefield/external/ExternalDistanceInteraction.hpp>
+#include <mjolnir/forcefield/external/PositionRestraintInteraction.hpp>
 #include <mjolnir/core/AxisAlignedPlane.hpp>
 #include <mjolnir/util/make_unique.hpp>
 #include <mjolnir/util/throw_exception.hpp>
