+++ mjolnir/input/read_global_interaction.hpp
@@ -1,10 +1,10 @@
 #ifndef MJOLNIR_INPUT_READ_GLOBAL_INTERACTION_HPP
 #define MJOLNIR_INPUT_READ_GLOBAL_INTERACTION_HPP
 #include <extlib/toml/toml.hpp>
-#include <mjolnir/interaction/global/GlobalPairInteraction.hpp>
-#include <mjolnir/interaction/global/GlobalPairLennardJonesInteraction.hpp>
-#include <mjolnir/interaction/global/GlobalPairUniformLennardJonesInteraction.hpp>
-#include <mjolnir/interaction/global/GlobalPairExcludedVolumeInteraction.hpp>
+#include <mjolnir/forcefield/global/GlobalPairInteraction.hpp>
+#include <mjolnir/forcefield/global/GlobalPairLennardJonesInteraction.hpp>
+#include <mjolnir/forcefield/global/GlobalPairUniformLennardJonesInteraction.hpp>
+#include <mjolnir/forcefield/global/GlobalPairExcludedVolumeInteraction.hpp>
 #include <mjolnir/forcefield/3SPN2/ThreeSPN2BaseBaseInteraction.hpp>
 #include <mjolnir/forcefield/PDNS/ProteinDNANonSpecificInteraction.hpp>
 #include <mjolnir/forcefield/PWMcos/PWMcosInteraction.hpp>
