+++ mjolnir/input/read_local_potential.hpp
@@ -3,17 +3,17 @@
 #include <extlib/toml/toml.hpp>
 #include <mjolnir/input/utility.hpp>
 
-#include <mjolnir/potential/local/HarmonicPotential.hpp>
-#include <mjolnir/potential/local/GoContactPotential.hpp>
-#include <mjolnir/potential/local/GoContactAttractivePotential.hpp>
-#include <mjolnir/potential/local/GoContactRepulsivePotential.hpp>
-#include <mjolnir/potential/local/ClementiDihedralPotential.hpp>
-#include <mjolnir/potential/local/GaussianPotential.hpp>
-#include <mjolnir/potential/local/PeriodicGaussianPotential.hpp>
-#include <mjolnir/potential/local/CosinePotential.hpp>
-#include <mjolnir/potential/local/SumLocalPotential.hpp>
-#include <mjolnir/potential/local/UniformPotential.hpp>
-#include <mjolnir/potential/local/WormLikeChainPotential.hpp>
+#include <mjolnir/forcefield/local/HarmonicPotential.hpp>
+#include <mjolnir/forcefield/local/GoContactPotential.hpp>
+#include <mjolnir/forcefield/local/GoContactAttractivePotential.hpp>
+#include <mjolnir/forcefield/local/GoContactRepulsivePotential.hpp>
+#include <mjolnir/forcefield/local/ClementiDihedralPotential.hpp>
+#include <mjolnir/forcefield/local/GaussianPotential.hpp>
+#include <mjolnir/forcefield/local/PeriodicGaussianPotential.hpp>
+#include <mjolnir/forcefield/local/CosinePotential.hpp>
+#include <mjolnir/forcefield/local/SumLocalPotential.hpp>
+#include <mjolnir/forcefield/local/UniformPotential.hpp>
+#include <mjolnir/forcefield/local/WormLikeChainPotential.hpp>
 #include <mjolnir/forcefield/FLP/FlexibleLocalAnglePotential.hpp>
 #include <mjolnir/forcefield/FLP/FlexibleLocalDihedralPotential.hpp>
 #include <mjolnir/forcefield/3SPN2/ThreeSPN2BondPotential.hpp>
